https://github.com/OpenPaaS-Suite/esn-frontend-inbox/issues/53: Moved inbox identities services to esn.inbox.libs
