+++ src/linagora.esn.unifiedinbox/app/services/jmap-helper/jmap-helper.js
@@ -2,7 +2,7 @@ const _ = require('lodash');
 require('../../../../esn.inbox.libs/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js');
 require('../../../../esn.inbox.libs/app/services/email-body/email-body.js');
 require('../with-jmap-client/with-jmap-client.js');
-require('../identities/inbox-identities.service.js');
+require('../../../../esn.inbox.libs/app/services/identities/inbox-identities.service.js');
 
 (function(angular) {
   'use strict';
